requirements.txt
python-telegram-bot[webhooks]==22.0
flask==3.1.0
requests
python-dotenv
